LayoutTests/imported/w3c/web-platform-tests/content-security-policy/style-src/stylenonce-allowed.sub-expected.txt

CONSOLE MESSAGE: Refused to apply a stylesheet because its hash, its nonce, or 'unsafe-inline' does not appear in the style-src directive of the Content Security Policy.
This text should be green.

This text should also be green.

Style correctly whitelisted via a 'nonce-*' expression in 'style-src' should be applied to the page.


Harness Error (TIMEOUT), message = null

NOTRUN Should fire securitypolicyviolation
PASS stylenonce-allowed
PASS stylenonce-allowed 1
